Seperate database management
Seperates out all the database related functions out of the
media_manager and also optimizing it more so it uses more
batching instead of single commits
